修复数值小数问题(我眼瞎了艹)

+++ src/main/java/io/github/codevine327/commandattributes/CommandImp.java
@@ -39,7 +39,7 @@ public class CommandImp implements CommandExecutor {
 
         String playerName = args[0];
         String statName = args[1];
-        int statValue = Integer.parseInt(args[2]);
+        double statValue = Double.parseDouble(args[2]);
 
         Player player = Bukkit.getPlayer(playerName);
