+++ src/main/java/io/github/codevine327/commandattributes/DataUtil.java
@@ -6,7 +6,7 @@ import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 
 public class DataUtil {
-    public static void addStat(Player player, String statName, int statValue, boolean writeFile) {
+    public static void addStat(Player player, String statName, double statValue, boolean writeFile) {
         PlayerData playerData = PlayerData.get(player.getUniqueId());
         playerData.getStats().getMap().getInstance(statName).addModifier(new StatModifier("cmdatt", statName, statValue));
 
@@ -15,7 +15,7 @@ public class DataUtil {
         }
     }
 
-    private static void writeData(Player player, String statName, int Value) {
+    private static void writeData(Player player, String statName, double Value) {
         FileConfiguration config = CommandAttributes.config;
         config.set(player.getUniqueId() + ".PLAYER_NAME", player.getName());
         config.set(player.getUniqueId() + "." + statName, Value);
